Updated Prometheus metrics.
Moved away from histograms, since I don't really know the
quantiles we support yet. It is probably likely that we
will eventually want to use histograms, until that time
we are going to be using summaries because they are less
overhead.

Moved from milliseconds to seconds, as that is the standard
practice that prometheus recommends. See more information
here: https://prometheus.io/docs/practices/naming/#base-units