Added metrics links to GUI.
Needed to expose a possible metrics URL. This does not need
to be set if no metrics are being monitored, but if it is set
it will show up in the about page.

I added the new defaults to the `dev_conf/config.yml`.

To make the about page fit all of this, I turned the Rest
documentation links into just a list of helpful links.