Updated queue deletion/clearing.
This update relies on setting the request status to canceled
via the ReST API. This has been causing us a couple of edge
case problems and sending the data back through the normal
data path should clear up a lot of things.

In particular, it is hard to keep gauges accurate without
this change.