Added Grafana Dashboards.
Changes to the docker-compose to include grafana and prometheus
by default. This enables our metrics to be scraped and the
default installation of beer-garden to be monitorable.

I added the two dashboards I created. You might notice they are
in two spots. There is a reason for this. The dashboards located
at `resources/grafana_dashboards` are the dashboards that get
uploaded to grafana.com.  These can be imported and when they
get imported by people, they get a prompt asking them to point
the dashboards at a prometheus datastore. Unfortunatley, you
cannot do the same for the "provisioned" dashboards that are
in the docker-compose section. In this section, I had to hard
code the datasource, which is okay because I'm also provisioning
the datasource, but still. That's why it is there.