Fixes #4
We had to change the default logging level from INFO to whatever
they set, or the system level. For the plugin runner, the
default logger (i.e. the one that bartender uses) should default
to whatever the system is set to.

However, the timestamp and unformatted loggers now use the level
specified by the plugin developer (INFO by default). The log
level is then set in the environment for the plugin. As a result
changes will need to happen in brewtils.