This solves beer-garden/beer-garden#79
This commit allows beer-garden to safely migrate from JSON config
files to YAML config files. There is one error case, when moving
the files around could cause application startup to fail. I believe
this is mostly mitigated because we perform the scary operation
last which should throw an error before catastrophic failure
occurrs.

We also guarantee that the file will be yaml. Since yaml is a
superset of JSON, the yaml parser can load JSON just fine in
the event that migrating/moving the YAML file does not complete
successfully.