Systems are now linked without their ID.
These changes only affect systems and their routes.

The URL routing changes are:

* /systems/{id} => /systems/{name}/{version}

A special note about versions though, the {version} will either
be a conrecete version or "latest" if it is the latest of the
systems currently loaded. This commit does not break backwards
compatibility for ID routing, and thus it is still in the
routes. However, no part of the UI should be using the ID for
systems anymore.

Views Tested:

* landing
* system ID
* System Admin
* Queue Admin