Swagger documentation and controller cleanup.
The APIs that exist should be in relatively their final form.
I still want to add PATCH support to the JobAPI, but I will
do that in a later commit. For now, the controller have all
been added and the OpenAPI documentation has been written.

I also added in the apscheduler as a dependency.

Finally, there is a new configuration value (web.host) which
was introduced to provide clarity on what address exactly we
are binding to. This will be important when scheduling jobs
because we want to use the EasyClient to hit our own API, but
it requires a host.