Added job success/error count tracking.
This required updating to the latest bg-utils with the success
and error counts on the job. Also, the job runner needs to
know the job id, so the scheduler was updated to add that.

Finally, the actually updating happens in the job_api. This is
not great from an encapsulation point-of-view, but we really
need to separate out more of the request processing from the
controllers.

The tests for the request_api were changed to actually test
the API itself. It slows down the tests, but since we dont
have very many, and it was way easier to test the new job
stuff, I changed it over to be a little bit more thorough since
it already caused an error earlier (see the missed .labels call
on the prometheus updates.)