Config loading on application initialization.
This included creating actions/reducers/action types
for the config calls. The actual API calls are in
actions, while the reducers handle the state. The
constants are defined in constants. This commit
also adds a nice error display and retry capability
if the API server is down for whatever reason. These
are found in the ErrorRetryDialog and RetryTimer
components.