Changed startup behavior for creating admin user.
We no longer create an anonymous user, or a plugin
user by default. This will instead be taken care
of as part of the application setup. In addition,
this change also allows us to store metadata on a
principal to set things like "Has the password
changed?" Finally, this also allows the default
password to be set in the environment so that the
default of "password" can be changed if a user
desires.