Moving to black as the code formatter.
We are keeping flake8 as of now. If we find that it has no use
then we can get rid of it. Currently, black fights with flake8
on at least 2 known problems, these have been added to the
ignore in the `setup.cfg`.

Black does not allow configuration from the `setup.cfg` and we
need configuration to ignore files, instead, black uses the new
`pyproject.toml` file. This was added to make sure we can
exclude the javascript source.