Improvements to getting user.
This fixes a bug where, if the cookie was set, but guest login
is enabled, we would authenticate you as a guest user, even
if the cookie was set properly for a real user. Now we ensure
that even if you are allowed to be an anonymous user, we check
the cookie to make sure you haven't already authenticated.