Sidebar routing working.
Sidebar routing now correctly routes to the appropriate view.
Lots of excess containers, but we will need them later. This
also meant a change for the way the layout is rendered. This
means that each of the containers dont have to worry about
rendering the sidebar like the system dashboar used to.