Adding authenticated routing.
These changes allow us to just use <AuthRoute /> where we would
normally use <Route /> to take care of logging in and such. Next
up will be an authView function of some kind that will display an
error if an authenticated person attempts to view a component they
do not have access to.