Made verison handler more resilient.
Before this change, if we could not acquire a thirft context
(i.e. if bartender was down) then we would error out on the
call to /version. Now we will simply say we dont know what
the bartender_version is but will continue to report the versions
we know without getting a thrift connection.